requirements.txt
Flask==3.1.0
PyMuPDF==1.25.5
python-dotenv
requests
gunicorn
